app/requirements.txt
asgiref==3.8.1
attrs==24.2.0
boto==2.49.0
boto3==1.19.9
botocore==1.22.12
certifi==2024.8.30
charset-normalizer==3.4.0
coreapi==2.3.3
coreschema==0.0.4
Django==4.2.5
django-cors-headers==4.2.0
django-filter==22.1
django-phonenumber-field==8.0.0
django-storages==1.12.3
djangorestframework==3.14.0
djangorestframework-datatables==0.7.0
dnspython==2.6.1
drf-spectacular==0.27.2
drf-writable-nested==0.7.0
drf-yasg==1.21.8
gunicorn==20.1.0
idna==3.10
inflection==0.5.1
itypes==1.2.0
Jinja2==3.1.4
jmespath==0.10.0
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
MarkupSafe==3.0.2
openapi-codec==1.3.2
packaging==24.1
phonenumbers==8.13.47
pillow==11.1.0
psycopg2-binary==2.9.10
python-dateutil==2.9.0.post0
pytz==2024.2
PyYAML==6.0.2
referencing==0.35.1
requests==2.32.3
rpds-py==0.20.0
s3transfer==0.5.2
simplejson==3.19.3
six==1.17.0
slack-sdk==3.21.3
sqlparse==0.5.1
typing_extensions==4.12.2
tzdata==2024.2
uritemplate==4.1.1
urllib3==1.26.20
whitenoise==6.2.0
